Created MongoDB "Hello World" function to test connections
Created a function for the default GET route in express that inserts a record with the data "Hello World" into the collection "Hello World" and database "Hello World". This tests connections to the remote MongoDB instance.

Installed MongoDB using npm, added it to the package.json file.

Connection string ("The uri constant") omitted from the commit.