Added the AddUserAccount function in the MongoConnection class
Added the MongoConnection class, with the constructor that takes a connection URI.

Added the AddUserAccount function, which takes JSON data, validates it with the Validator ValidateUserAccount function, hashes the password with the bcrypt package and then inserts the record into the MongoDB database.

Installed bcrypt using the node package manager.