Added error-handling functionality to the login page
Added additional functionality to the "/login" route that appends a new "authfail=true" parameter on an unsuccessful login.
Added the HandleAuthenticationFailure function to the new login.js file, which produces an alert if the authfail parameter in the URL is present and set to true.
Added references to the new JavaScript files in the login page, as well as a JQuery CDN.