Added generation and rendering of an instanced map grid
Added the MapScreen class.
Added the InitialiseScene method to MapScreen. Sets up the rendering environment.
Added the CreateNewMap method to MapScreen. Creates a new Map instance and then generates instanced geometry and colour buffers for rendering.
Added the ResizeDisplay method to MapScreen. Changes the rendering resolution and camera aspect if the canvas is resized.
Added the Render and BeginRendering methods to MapScreen. These are responsible for maintaining the main render loop.