Created the InstancedObjectPicker class and associated methods
Created the ClearObjects method in InstancedObjectPicker. Clears stored objects from the previous frame.
Created the PickClickedObject method in InstancedObjectPicker. Selects an instance of an object on a mouse click, adding, removing or selecting an object depending on the value of activeSelectType.
Created the PickHoveredObject method in InstancedObjectPicker. Selects an instance of an object on a mouse hover, creating a 3D outline around the selected object.
Created the AddBlockToGrid method in InstancedObjectPicker. Adds a value to the specified entries on the map.heightMap property, then re-renders the corresponding blocks on the rendered map.
Created the RemoveBlockFromGrid method in InstancedObjectPicker. Subtracts a value from the specified entries on the map.heightMap property, then re-renders the corresponding blocks on the rendered map.
Created the SelectBlockOnGrid method in InstancedObjectPicker. Creates a HTML element at the specified location.