Added the HTMLGenerator class
Added the RemoveLabels method in HTMLGenerator. Helper method to remove the active method. Called when a new label is added.
Added the AddLabel method in HTMLGenerator. Adds a new label, linked to a specific instance of an object.
Added the MoveLabel method in HTMLGenerator. Moves the active label when the linked object instance is moved.