Created the SetHeight and IncreaseHeightOfBlock functions to set block height and re-render
Created the SetHeight function in the Map class to set the height value on the heightMap matrix if the value is a positive value above 1.
Created the IncreaseHeightOfBlock function in the MapScreen class, setting the block height and re-rendering the chosen block at the correct height.
Removed height modification code from the heightModFunction in the HTMLGenerator.