Fixed issue where blocks could incorrectly be rendered at negative height
Modified SetHeight so that it returns the assigned height value, allowing the value to be set and read in a single call.
Modified SetHeight so that it sets if the value is above 0, allowing the height to be set to 1.
Modified IncreaseHeightOfBlock to make use of the returned value so the block is only rendered to a correctly set height.