Added updated map loading and saving functionality with correct separation of functionality
Moved networking and file loading code to the Map class, which is now responsible for its own file loading.
Added promise functionality to the LoadMap method in the Map class, to ensure it can be performed asynchronously without issue.
Added the SaveMap function to the Map class, which makes a PUT request asynchronously.
Added an event listener to trigger the save function in the Map class.