Added functionality to generate an empty character matrix, as well as to add and retrieve character records to the Map class
Added GenerateCharacterMatrix, which generates an empty matrix for storing character records.
Added AddCharacter, which takes X and Y coordinates as parameters and constructs a new Character instance, storing in the corresponding X and Y indices in the characterMatrix property.
Added GetCharacter, which takes X and Y coordinates as parameters and retrieves from the corresponding indices in the characterMatrix property.