Added the ability to add and remove characters to the map, as well as modifying descriptive details
Added the AddCharacter method to the InstancedObjectPicker class. The method adds a new character record to the map instance, creates a new cylinder mesh and adds a new label using the HTMLManager class.
Added the AddCharacterLabel method to the HTMLManager class, which adds a new label with character specific elements, as well as adding event handlers to manage user input.
Added functionality to the MapRender class to produce new meshes for characters stored in the map file.