Refactored the AddBlockToGrid method in the InstancedObjectPicker class into the MapScreen class
Refactored the AddBlockToGrid method in the InstancedObjectPicker class. Functionality is now in the AddBlock method in the MapScreen class which is called by an event handler. The AddBlockToGrid method now creates a new event and dispatches it to the document.
Added the creation of an event listener in the InitialiseScene method in MapScreen