Modified the AddBlock method to add height subtraction functionality
The AddBlock method in the MapScreen class is now the ChangeBlock method, with addition/subtraction occuring based on the activeSelectType value.
The AddBlock method in the Map class has been replaced with the AddToHeight method, which now checks if the value is less than 1 and allows for the addition of negative and varying values, as opposed to 1.