Created modal generation and management functionality
Created the NewMapModal, AlertModal and ProcessModal, which are used to generate and manage modal HTML elements for specific alert messages, forms and loading messages.
Implemented the GenerateHTML methods in all three modal classes. GenerateHTML creates and appends DOM elements for the modals to the document.
Implemented the Show methods in all three modal classes, which reveal the modal.
Implemented the Hide methods in all three modal classes, which hide the modal.
Implemented the Submit method in the NewMapModal, which parses form values and creates a new map using them.