Modified Client-side scripts to prepare for socket.io functionality
Moved the InitialiseEventListeners function into the MapScreen class. Modified to bind the current object scope to events.
Added the DetermineSessionType method to MapScreen. Reads the SessionType cookie to determine the chosen type of session and calls the appropriate initialisation method.
Added the HostSessionInitialise, ClientSessionInitialise and EditSessionInitialise method prototypes.
Added a new button to the DOM generation in the PopulateMapList function.
Added an event handler for the new button.
Added functionality to set the SessionType cookie in the event handlers for the editButton and the hostButton in PopulateMapList.