Implemented 2-way map modification functionality and user account retrieval
Implemented the RenderScene method in the MapScreen class. The method handles the generation of a new scene and map grid. Is used to refresh the page when changes to the map are detected.
Modified calls to the AddCharacterLabel method to determine if the user is allowed to modify a character.