Added back end methods for interacting with hidden region functionality
Added the ToggleRegionVisibility event handler which triggers the RevealRegion method in the Map class.
Added the AddNewRegion event handler which triggers the AddNewRegion method in the Map class.
Added the AddBlockToHiddenRegion event handler which triggers the AddBlockToHiddenRegion method in the Map class.
Added the RemoveBlockFromHiddenRegion event handler which triggers the RemoveBlockFromHiddenRegion method in the Map class.
Added code to register new event listeners to trigger the new event handlers.