Modified hidden areas to make them less obvious to players.
Modified functionality. Block scale is no longer set to 0 as that produced very obvious gaps in the grid. Now the dimensions of the block are set to the default state, displaying as flat ground.
Added functionality to display the hidden blocks button as selected when clicked on.
Fixed incorrect method calls.